Move setup into middleware
