+++ server/middleware/configureEstablishment.js
@@ -1,12 +1,24 @@
+const { v4: uuid } = require('uuid');
+const { pathOr } = require('ramda');
 const {
+  getEstablishmentId,
   getEstablishmentFormattedName,
   getEstablishmentPrefix,
 } = require('../utils');
 
 const configureEstablishment = () => (req, res, next) => {
-  req.session.prison = req.session.establishmentName;
+  if (req.session && (!req.session.id || !req.session.establishmentId)) {
+    const replaceUrl = /-prisoner-content-hub.*$/g;
+    const establishmentName = pathOr('wayland', ['headers', 'host'], req)
+      .split('.')[0]
+      .replace(replaceUrl, '');
 
-  res.locals.establishmentId = req.session.establishmentId;
+    req.session.id = uuid();
+    req.session.establishmentName = establishmentName;
+    req.session.establishmentId = getEstablishmentId(establishmentName);
+  }
+
+  res.locals.feedbackId = uuid();
   res.locals.establishmentDisplayName = `${getEstablishmentPrefix(
     req.session.establishmentId,
   )} ${getEstablishmentFormattedName(req.session.establishmentId)}`;
